+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Config and Connection APIs turned from class-level to instance-level
 - configuration is not set at the RemoteResource subclass level, giving full control over payload data type and authentication concerns
 - groundwork laid for easily adding additional supported configurations
+- custom method support
 
 ###[0.2.0] - 2015-05-01
 - php namespace pattern applied